change servo angles from i32 to f32

+++ ruspy/src/drive.rs
@@ -121,10 +121,10 @@ impl Servo {
         Ok(())
     }
 
-    pub fn angle(&mut self, angle: i32) {
-        let angle = angle.clamp(-90, 90);
-        let pw_time = map_range((-90, 90), (MIN_PW.into(), MAX_PW.into()), angle);
-        let pw_time = pw_time.clamp(MIN_PW.into(), MAX_PW.into());
+    pub fn angle(&mut self, angle: f32) {
+        let angle = angle.clamp(-90.0, 90.0);
+        let pw_time = map_range((-90.0, 90.0), (MIN_PW.into(), MAX_PW.into()), angle);
+        let pw_time = pw_time.clamp(MIN_PW.into(), MAX_PW.into()) as i32;
         let _ = self.pulse_width_time(pw_time);
     }
 }
